+++ bin/utils.js
@@ -161,11 +161,26 @@ export const getPluginDiffsForSanity = (pluginLookup, plugins) => {
  * @returns {BuildPluginEntity}
  */
 const convertCmsChangesToRepoPlugin = (plugin) => {
-  const { compatibility, description, packageName, status: rawStatus, repoUrl: repo, title: name, version } = plugin
+  const {
+    compatibility,
+    description,
+    packageName,
+    status: rawStatus,
+    repoUrl: repo,
+    title: name,
+    version,
+    metadata,
+  } = plugin
   /**
    * @type {BuildPluginEntity['status']}
    */
   const status = rawStatus === 'deactivated' ? rawStatus.toUpperCase() : undefined
+
+  const variables =
+    metadata?.variables.map(({ name: varName, description: varDescription }) => ({
+      name: varName,
+      description: varDescription,
+    })) || []
 
   return {
     name,
@@ -175,6 +190,7 @@ const convertCmsChangesToRepoPlugin = (plugin) => {
     version,
     status,
     compatibility,
+    variables,
   }
 }
 
@@ -189,6 +205,11 @@ const stripNullifiedFields = (plugin) => {
   if (!plugin.compatibility) {
     // eslint-disable-next-line no-param-reassign
     delete plugin.compatibility
+  }
+
+  if (!plugin.variables || plugin.variables.length === 0) {
+    // eslint-disable-next-line no-param-reassign
+    delete plugin.variables
   }
 
   return plugin